Group all trace-recorder related classes (but not trace data classes) into separate package.
It allows to extract it as separate module in the future.
