+++ src/jvm/main/org/jetbrains/kotlinx/lincheck/tracerecorder/agent/TraceRecorderInjections.kt
@@ -8,12 +8,12 @@
  * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.jetbrains.kotlinx.lincheck.traceagent
+package org.jetbrains.kotlinx.lincheck.tracerecorder.agent
 
-import org.jetbrains.kotlinx.lincheck.strategy.tracerecorder.TraceRecorder
+import org.jetbrains.kotlinx.lincheck.tracerecorder.TraceRecorder
+import org.jetbrains.kotlinx.lincheck.traceagent.TraceAgentParameters
 import org.jetbrains.kotlinx.lincheck.transformation.InstrumentationMode
 import org.jetbrains.kotlinx.lincheck.transformation.LincheckJavaAgent
-import org.objectweb.asm.commons.Method
 
 internal object TraceRecorderInjections {
