+++ src/jvm/main/org/jetbrains/kotlinx/lincheck/tracerecorder/agent/TraceRecorderMethodTransformer.kt
@@ -8,10 +8,9 @@
  * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.jetbrains.kotlinx.lincheck.traceagent
+package org.jetbrains.kotlinx.lincheck.tracerecorder.agent
 
 import org.jetbrains.kotlinx.lincheck.transformation.ASM_API
-import org.jetbrains.kotlinx.lincheck.transformation.invokeInsideIgnoredSection
 import org.jetbrains.kotlinx.lincheck.transformation.invokeStatic
 import org.objectweb.asm.commons.AdviceAdapter
 import org.objectweb.asm.commons.GeneratorAdapter
