+++ src/jvm/main/org/jetbrains/kotlinx/lincheck/tracerecorder/transformers/MethodCallMinimalTransformer.kt
@@ -1,18 +1,19 @@
 /*
  * Lincheck
  *
- * Copyright (C) 2019 - 2024 JetBrains s.r.o.
+ * Copyright (C) 2019 - 2025 JetBrains s.r.o.
  *
  * This Source Code Form is subject to the terms of the
  * Mozilla Public License, v. 2.0. If a copy of the MPL was not distributed
  * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.jetbrains.kotlinx.lincheck.transformation.transformers
+package org.jetbrains.kotlinx.lincheck.tracerecorder.transformers
 
 import org.jetbrains.kotlinx.lincheck.transformation.*
 import org.jetbrains.kotlinx.lincheck.tracedata.MethodDescriptor
 import org.jetbrains.kotlinx.lincheck.tracedata.methodCache
+import org.jetbrains.kotlinx.lincheck.transformation.transformers.MethodCallTransformerBase
 import org.objectweb.asm.Opcodes.*
 import org.objectweb.asm.Type.*
 import org.objectweb.asm.commons.*
