+++ src/jvm/main/org/jetbrains/kotlinx/lincheck/tracerecorder/transformers/ObjectCreationMinimalTransformer.kt
@@ -8,7 +8,7 @@
  * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.jetbrains.kotlinx.lincheck.transformation.transformers
+package org.jetbrains.kotlinx.lincheck.tracerecorder.transformers
 
 import sun.nio.ch.lincheck.*
 import org.jetbrains.kotlinx.lincheck.transformation.*
