+++ src/jvm/main/org/jetbrains/kotlinx/lincheck/transformation/LincheckClassVisitor.kt
@@ -14,6 +14,8 @@ import org.jetbrains.kotlinx.lincheck.*
 import org.jetbrains.kotlinx.lincheck.traceagent.isInTraceDebuggerMode
 import org.jetbrains.kotlinx.lincheck.tracedata.CodeLocations
 import org.jetbrains.kotlinx.lincheck.tracedata.FinalFields
+import org.jetbrains.kotlinx.lincheck.tracerecorder.transformers.MethodCallMinimalTransformer
+import org.jetbrains.kotlinx.lincheck.tracerecorder.transformers.ObjectCreationMinimalTransformer
 import org.objectweb.asm.*
 import org.objectweb.asm.Opcodes.*
 import org.objectweb.asm.commons.*
